feat(coderabbit): 코드래빗 설정

+++ src/main/java/project/freemates2/api/university/domain/University.java
@@ -27,9 +27,9 @@ public class University {
 
   private String slug;
 
-  private int lat; // 위도
+  private Double lat; // 위도
 
-  private int lng; // 경도
+  private Double lng; // 경도
 
   private String iconUrl;
